+++ src/components/OrderRight/components/SelectBoxQuantity.scss
@@ -1,0 +1,66 @@
+@import '../../../styles/variables.scss';
+
+.selectBoxQuantity {
+  position: relative;
+
+  .selectBox {
+    display: flex;
+    justify-content: space-between;
+    width: 50px;
+    padding: 5px 11px;
+    border: 1px solid $GRAY333;
+    background: $bgColor;
+    color: $GRAY333;
+
+    cursor: pointer;
+  }
+
+  .selectArrow {
+    position: absolute;
+    top: 3px;
+    right: 67px;
+    width: 9px;
+
+    img {
+      width: 100%;
+      padding-bottom: 1px;
+
+      cursor: pointer;
+    }
+  }
+
+  ul {
+    display: none;
+    width: 48px;
+    height: 60px;
+    padding: 3px 0;
+    margin-top: 2px;
+    border: 1px solid $GRAY333;
+    background: $bgColor;
+
+    overflow: scroll;
+
+    li {
+      button {
+        width: 48px;
+        padding: 5px 15px;
+        border: none;
+        background: $bgColor;
+
+        cursor: pointer;
+      }
+    }
+
+    &.show {
+      display: block;
+      width: 48px;
+      height: 60px;
+      margin-top: 2px;
+      border: 1px solid $GRAY333;
+    }
+
+    &::-webkit-scrollbar {
+      display: none;
+    }
+  }
+}
